+++ vscode/src/main.ts
@@ -1,5 +1,5 @@
 import * as vscode from 'vscode';
-import { createService } from './searchService';
+import { createService as createSearchService } from './searchService';
 import { VsCode } from './ui/vscode';
 import { NoteSearcher } from './noteSearcher';
 import { DeadLinkFinder } from './DeadLinkFinder';
@@ -10,31 +10,31 @@ export const extensionId = 'uozuaho.note-searcher';
 
 export function activate(context: vscode.ExtensionContext) {
   const ui = new VsCode();
-  const searcher = createService(extensionDir()!);
+  const searcher = createSearchService(extensionDir()!);
   const deadLinkFinder = new DeadLinkFinder(createFileSystem());
-  const configProvider = new NoteSearcherConfigProvider();
-  const noteSearcher = new NoteSearcher(ui, searcher, deadLinkFinder, configProvider);
+  const configProvider = new NoteSearcherConfigProvider(context);
 
-  const search = vscode.commands.registerCommand(
-    'noteSearcher.search', async () => await noteSearcher.search());
-
-  const index = vscode.commands.registerCommand(
-    'noteSearcher.index', async () => await noteSearcher.index());
-
-  const openFile = vscode.commands.registerCommand(
-    'noteSearcher.searchResults.openFile',
-    file => vscode.window.showTextDocument(file));
-
-  const docChangeHandler = ui.createOnDidChangeTextDocumentHandler();
-  const docSaveHandler = ui.createOnDidSaveDocumentHandler();
+  const noteSearcher = new NoteSearcher(ui,
+    searcher, deadLinkFinder, configProvider);
 
   context.subscriptions.push(
-    search,
-    index,
-    openFile,
-    docChangeHandler,
-    docSaveHandler
+    vscode.commands.registerCommand(
+      'noteSearcher.search', async () => await noteSearcher.search()),
+    vscode.commands.registerCommand(
+      'noteSearcher.index', async () => await noteSearcher.index()),
+    vscode.commands.registerCommand(
+      'noteSearcher.searchResults.openFile',
+      file => vscode.window.showTextDocument(file)),
+    vscode.commands.registerCommand(
+      'noteSearcher.enableCurrentDir', () => noteSearcher.enable()),
+    vscode.commands.registerCommand(
+      'noteSearcher.disableCurrentDir', () => noteSearcher.disable()),
+
+    ui.createOnDidChangeTextDocumentHandler(),
+    ui.createOnDidSaveDocumentHandler()
   );
+
+  noteSearcher.notifyExtensionActivated();
 }
 
 export function deactivate() {}